requirements.txt
pandas
openpyxl
numpy
faiss-cpu
sentence-transformers
openai
python-dotenv
requests
tqdm
regex
pytest
pypdf
